Posts

+++ routes/posts.js
@@ -1,7 +1,12 @@
 /* For video/image posting routes */
 
 const express = require("express")
-const { pgQuery, s3Upload, s3Retrieve, s3Delete } = require('../functions/general_functions')
+const { pgQuery, s3Upload, s3Retrieve, s3Delete, requestLimiter } = require('../functions/general_functions')
+const { validationResult } = require('express-validator')
+const { validatePostVideo, validateGetPost, validateParamId, validateGetCategoryPost, validateGetPosts } = require('../functions/validators/posts_validators')
+
+const pool = require("../pgdb")
+
 const multer = require('multer')
 const { getItemPrimaryKey, getItemPartitionKey, putItem, deleteItem } = require('../functions/dynamoDB_functions');
 const { setPostStats } = require("../dynamo_schemas/dynamo_schemas")
@@ -20,6 +25,7 @@ router.get("/testing", async (req, res) => {
     console.error(err.message)
   }
 })
+ 
 
 /* Functions for Posts */
 /* returns the total likes and views per post */
@@ -34,7 +40,6 @@ async function getPostStats(postId){
 
   const stats = await getItemPartitionKey(params)
 
-  console.log(stats)
   return stats[0]
 
 } 
@@ -74,43 +79,79 @@ async function checkLike(postId, userId) {
 
 
 /* Posting a post to the database */
-router.post("/postvideo", upload.any(), async (req, res) => {
+router.post("/posts/:id", requestLimiter, upload.any(), validatePostVideo(), async (req, res, next) => {
   try {
-    const { user_id, post_title, post_description, category_id } = req.body
+
+    const errors = validationResult(req)
+  
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
+    const userId = parseInt(req.params.id)
+    const { post_title, post_description, category_id_list, hashtag_id_list, recipe_description, recipe_ingredients, recipe_equipment, recipe_steps, recipe_preparation_time, recipe_serving } = req.body
 
     //Used to upload to s3 bucket
-    const newVideoName = await s3Upload(req.files[0])
-    const newThumbNaileName = await s3Upload(req.files[1])
+    const [newVideoName, newThumbNaileName] = await Promise.all([
+      s3Upload(req.files[0]),
+      s3Upload(req.files[1])
+    ])
 
-    const newPost = await pgQuery(
-      `INSERT INTO posts (user_id, post_title, post_description, video_name, thumbnail_name, category_id, created_at, updated_at) 
-      VALUES ($1, $2, $3, $4, $5, $6, NOW(), NOW()) RETURNING *`,
-      user_id, post_title, post_description, newVideoName, newThumbNaileName, category_id
-    )
+    const client = await pool.connect();
 
-    //INSERT RECIPE aswell
+    try {
+      await client.query('BEGIN')
 
-    const { post_id } = newPost.rows[0]
-    
-    const postStatsSchema = setPostStats(post_id)
+      const insertRecipeQuery = 'INSERT INTO recipes (recipe_description, recipe_ingredients, recipe_equipment, recipe_steps, preparation_time, recipe_servings) VALUES ($1, $2, $3, $4, $5, $6) RETURNING recipe_id';
+      const recipeValues = [recipe_description, JSON.parse(recipe_ingredients), JSON.parse(recipe_equipment), JSON.parse(recipe_steps), recipe_preparation_time, recipe_serving];
+      const { rows } = await client.query(insertRecipeQuery, recipeValues);
 
-		await putItem("Post_Stats", postStatsSchema)
-    
-    console.log("Video Posted")
-    res.json(newPost.rows[0])
+      const { recipe_id } = rows[0];
+
+      const insertPostQuery = 'INSERT INTO posts (user_id, post_title, post_description, video_name, thumbnail_name, category_id_list, hashtag_id_list, recipe_id, post_created_at, post_updated_at) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, NOW(), NOW()) RETURNING *';
+      const postValues = [userId, post_title, post_description, newVideoName, newThumbNaileName, JSON.parse(category_id_list), JSON.parse(hashtag_id_list), recipe_id];
+      const newPost = await client.query(insertPostQuery, postValues)
+      
+      const { post_id } = newPost.rows[0]
+
+      const updatePostQuery = 'UPDATE recipes SET post_id = $1 WHERE recipe_id = $2';
+      const postUpdateValues = [post_id, recipe_id];
+      await client.query(updatePostQuery, postUpdateValues)
+
+      const postStatsSchema = setPostStats(post_id)
+      await putItem("Post_Stats", postStatsSchema)
+
+      console.log("Video Posted" + post_id)
+      await client.query('COMMIT')
+
+      res.json(newPost.rows[0])
+    } catch (err) {
+      await client.query('ROLLBACK')
+      next(err)
+    } finally {
+      client.release()
+    }
   } catch (err) {
-    console.error(err.message)
+    next(err)
   }
 })
+
 
 /* Get a specific post and its category
   /api/posts/getpost/8?user_id=3
 */
-router.get("/getpost/:id", async (req, res) => {
+router.get("/posts/:id", requestLimiter, validateGetPost(), async (req, res, next) => {
   try {
 
-    const postId = req.params.id
+    const errors = validationResult(req)
+  
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
+    const postId = parseInt(req.params.id)
     
+    //NEED TO UPDATE THIS TO THE NEW CATEGORY SYSTEM
     const specificPost = await pgQuery(`
       SELECT posts.*, users.*, categories.*, posts.created_at AS post_created_at, posts.updated_at AS post_updated_at 
       FROM posts JOIN users ON posts.user_id = users.user_id JOIN categories ON posts.category_id = categories.category_id 
@@ -118,15 +159,19 @@ router.get("/getpost/:id", async (req, res) => {
       , postId
     )
 
+    if (!specificPost.rows[0]){ res.status(400).json({ error: `No Post With Id Of ${postId}`})}
+
     const { post_id, user_id, post_title, post_description, video_name, thumbnail_name, name, category_id,
     username, profile_picture, post_created_at, post_updated_at } = specificPost.rows[0]
 
-    const videoUrl = await s3Retrieve(video_name)
-    const thumbnailUrl = await s3Retrieve(thumbnail_name)
+    const [videoUrl, thumbnailUrl, postStats, liked] = await Promise.all([
+      s3Retrieve(video_name),
+      s3Retrieve(thumbnail_name),
+      getPostStats(postId),
+      checkLike(postId, parseInt(req.query.user_id))
+    ])
 
-    const { view_count, like_count, comments_count } = await getPostStats(parseInt(postId))
-
-    const liked = await checkLike(parseInt(postId), parseInt(req.query.user_id))
+    const { view_count, like_count, comments_count } = postStats
     
     const responseData = {
       post_id,
@@ -149,44 +194,62 @@ router.get("/getpost/:id", async (req, res) => {
 
     res.json(responseData)
   } catch (err) {
-    console.error(err.message)
+    next(err)
   }
 })
 
 
 /* Deletes a specific post */
-router.delete("/deletepost/:id", async (req, res) => {
+router.delete("/posts/:id", requestLimiter, validateParamId(), async (req, res, next) => {
   try {
+
+    const errors = validationResult(req)
+  
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
 
     const postId = req.params.id
 
-    const post = await pgQuery(`SELECT * FROM posts WHERE post_id=$1`, postId)
+    const post = await pgQuery(`SELECT * FROM posts WHERE post_id = $1`, postId)
 
-    await s3Delete(post.rows[0].video_name)
-    await s3Delete(post.rows[0].thumbnail_name)
+    const { video_name, thumbnail_name } = post.rows[0]
 
-    await pgQuery(`DELETE FROM posts WHERE post_id = $1`, postId)
 
-    await removeLikesViews(parseInt(postId))
+    //CHANGE THIS SO THAT IF ONE DOESNT HAPPEN NONE DO
+    //MAYBE CREATE A DELETE FOLDER IN S3. To allow rollback
+    await Promise.all([
+      s3Delete(video_name),
+      s3Delete(thumbnail_name),
+      pgQuery(`DELETE FROM posts WHERE post_id = $1`, postId),
+      removeLikesViews(parseInt(postId))
+    ])
 
     console.log("Post Deleted")
     res.json({"Status": "Image Deleted"})
   } catch (err) {
-    console.error(err.message)
+    next(err)
   }
 })
 
-
 /* Get 15 random posts */
-router.get("/homepage/getposts", async (req, res) => {
+router.get("/homepage/posts", requestLimiter, validateGetPosts(), async (req, res, next) => {
   try {
 
+    const errors = validationResult(req)
+  
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
+    //THIS NEEDS TO CHANGE TO CORRESPOND TO THE NEW CATEGORY SYSTEM.
+    //NEEDS TO ADD PAGINATION AND CACHHING
     const randomPosts = await pgQuery(`
-      SELECT posts.*, users.*, categories.*, posts.created_at AS post_created_at, posts.updated_at AS post_updated_at 
-      FROM posts JOIN users ON posts.user_id = users.user_id JOIN categories ON posts.category_id = categories.category_id 
+      SELECT posts.*, users.*, categories.*
+      FROM posts JOIN users ON posts.user_id = users.user_id JOIN categories ON posts.category_id = categories.category_id
       ORDER BY RANDOM() LIMIT 15;
     `)
-
+    
     const processedPosts = await Promise.all(
       randomPosts.rows.map(async (post) => {
 
@@ -202,45 +265,61 @@ router.get("/homepage/getposts", async (req, res) => {
         return { ...rest, video_url: videoUrl, thumbnail_url: thumbnailUrl, view_count, like_count, liked }
       })
     )
-      
+    
     res.json({"posts": processedPosts})
 
   } catch (err) {
-    console.error(err.message)
+    next(err)
   }
 })
 
 /* Getting Ingredient For Specific Post */
-router.get("/getrecipe/:id", async (req, res) => {
+router.get("/posts/recipe/:id", requestLimiter, validateParamId(), async (req, res, next) => {
   try {
+
+    const errors = validationResult(req)
+  
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
     const recipeId = req.params.id
     const specificRecepie = await pgQuery(`SELECT * FROM recipes WHERE recipe_id = $1`, recipeId)
 
     res.json({"recipe": specificRecepie.rows[0]})
 
   } catch (err) {
-    console.error(err.message)
+    next(err)
   }
 })
 
 
 /* Posting For Sending Video To Friend */
-router.post("/homepage/postsendvideo", async (req, res) => {
+router.post("/posts/sendvideo", requestLimiter,  async (req, res, next) => {
   try {
     //Will need to update the messages table. and update a sent table.
 
  
   } catch (err) {
-    console.error(err.message)
+    next(err)
   }
 })
 
 
 /* Get 15 random posts for a specific category */
-router.get("/discover/getcategoryposts/:category", async (req, res) => {
+router.get("/categoryposts/:id", requestLimiter, validateGetCategoryPost(), async (req, res, next) => {
   try {
-    const categoryId = req.params.category
 
+    const errors = validationResult(req)
+  
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
+    //REQUIRES PAGINATION AND CACHING
+    const categoryId = req.params.id
+
+    // CHANGE THIS TO MATCH THE NEW CATEGORY SYSTEM
     const randomPosts = await pgQuery(`
       SELECT posts.*, users.*, categories.*, posts.created_at AS post_created_at, posts.updated_at AS post_updated_at 
       FROM posts 
@@ -269,7 +348,7 @@ router.get("/discover/getcategoryposts/:category", async (req, res) => {
     res.json({"posts": processedPosts})
 
   } catch (err) {
-    console.error(err.message)
+    next(err)
   }
 })
